+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
 * Added our own constructors for each of the Protobuf messages, so that when you add a new field it's harder to forget to add it
+* Added a `KurtosisApiVersion` value to Go which contains the version of this library, which is updated automatically during the release process of this repo
 
 ### Fixes
 * Fixed a bug where `stacktrace.Propagate` was incorrectly used instead of `stacktrace.NewError`